+++ CHANGELOG.md
@@ -69,6 +69,7 @@
 - PIM-10578: Fix the search on system filters in the rule engine UI
 - PIM-10592: Fix DQI score showing "In Progress" - data_quality_insights_evaluations returns an error
 - PIM-10619: Avoiding checks on expiration dates of certificates when disabling SSO
+- PIM-10621: Fix API returning an empty list instead of an empty object for published products
 
 ## Improvements